simplify include directory
