+++ src/usr/Makefile.am
@@ -59,9 +59,7 @@ libxio_includedir = $(includedir)/
 
 # the list of header files that belong to the library (to be installed later)
 libxio_include_HEADERS = $(top_srcdir)/include/libxio.h  	\
-			 $(top_srcdir)/include/xio_base.h	\
-			 $(top_srcdir)/include/xio_user.h	\
-			 $(top_srcdir)/include/xio_predefs.h
+			 $(top_srcdir)/include/xio_base.h
 
 
 libxio_headers = 	./xio/get_clock.h 			\
